+++ src/assets/metadata/index.js
@@ -1,0 +1,5 @@
+import aws from 'src/assets/metadata/aws.json';
+
+export default {
+  aws,
+};